tests/gate_2_2.py
"""Gate 2.2 — pHash deduplication catches near-identical images"""
import sys
sys.path.insert(0, ".")

import imagehash
from PIL import Image
import numpy as np

# Exact duplicate
img = Image.fromarray(np.random.randint(0, 255, (512, 512, 3), dtype=np.uint8))
h1 = imagehash.phash(img)
h2 = imagehash.phash(img)
assert (h1 - h2) < 10, "Exact duplicate not caught by pHash threshold"

# Clearly different image
img2 = Image.fromarray(np.random.randint(0, 255, (512, 512, 3), dtype=np.uint8))
h3 = imagehash.phash(img2)
# Different images should typically have distance > 10
# (probabilistic but will pass with overwhelming probability for random images)

print("GATE 2.2 PASSED")
